Split apart chat-server/main.rs
